+++ src/rpc/ddl_apply.rs
@@ -1,7 +1,8 @@
 use crate::op::Ddl;
 use crate::storage::Clusterwide;
 use crate::storage::{
-    ddl_create_function_on_master, ddl_create_space_on_master, ddl_drop_space_on_master,
+    ddl_create_function_on_master, ddl_create_space_on_master, ddl_drop_function_on_master,
+    ddl_drop_space_on_master,
 };
 use crate::storage::{local_schema_version, set_local_schema_version};
 use crate::tlog;
@@ -139,6 +140,17 @@ pub fn apply_schema_change(
 
         Ddl::CreateProcedure { id, ref name, .. } => {
             if let Err(e) = ddl_create_function_on_master(id, name) {
+                return Err(Error::Aborted(e.to_string()));
+            }
+        }
+
+        Ddl::DropProcedure { id, .. } => {
+            if !is_commit {
+                return Ok(());
+            }
+
+            let abort_reason = ddl_drop_function_on_master(id).map_err(Error::Other)?;
+            if let Some(e) = abort_reason {
                 return Err(Error::Aborted(e.to_string()));
             }
         }