+++ CHANGELOG.md
@@ -25,6 +25,8 @@ with the `YY.MINOR.MICRO` scheme.
 
 - Clusterwide SQL supports procedure creation.
 
+- Clusterwide SQL supports procedure deletion.
+
 - Make Proc API the main picodata API. Lua API is considered deprecated.
 
 - New "pico_service" builtin user is used for all inter-instance system